backend/requirements.txt
fastapi
pydantic
requests
stripe
python-dotenv
